requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==20.3.0
    # via jsonschema
awsebcli==3.19.2
    # via -r requirements.in
bcrypt==3.2.0
    # via paramiko
blessed==1.17.12
    # via awsebcli
boto3==1.16.49
    # via -r requirements.in
botocore==1.19.49
    # via
    #   awsebcli
    #   boto3
    #   s3transfer
cached-property==1.5.2
    # via docker-compose
cement==2.8.2
    # via awsebcli
certifi==2020.12.5
    # via requests
cffi==1.14.4
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
chardet==3.0.4
    # via requests
colorama==0.4.3
    # via awsebcli
cryptography==3.3.1
    # via paramiko
docker-compose==1.25.5
    # via awsebcli
docker[ssh]==4.4.1
    # via docker-compose
dockerpty==0.4.1
    # via docker-compose
docopt==0.6.2
    # via docker-compose
fire==0.3.1
    # via -r requirements.in
future==0.16.0
    # via awsebcli
idna==2.10
    # via requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via docker-compose
paramiko==2.7.2
    # via docker
pathspec==0.5.9
    # via awsebcli
pycparser==2.20
    # via cffi
pynacl==1.4.0
    # via paramiko
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   awsebcli
    #   botocore
pyyaml==5.3.1
    # via
    #   awsebcli
    #   docker-compose
requests==2.24.0
    # via
    #   awsebcli
    #   docker
    #   docker-compose
s3transfer==0.3.3
    # via boto3
semantic_version==2.5.0
    # via awsebcli
six==1.15.0
    # via
    #   awsebcli
    #   bcrypt
    #   blessed
    #   cryptography
    #   docker
    #   docker-compose
    #   dockerpty
    #   fire
    #   jsonschema
    #   pynacl
    #   python-dateutil
    #   websocket-client
termcolor==1.1.0
    # via
    #   awsebcli
    #   fire
texttable==1.6.3
    # via docker-compose
urllib3==1.25.11
    # via
    #   awsebcli
    #   botocore
    #   requests
wcwidth==0.1.9
    # via
    #   awsebcli
    #   blessed
websocket-client==0.57.0
    # via
    #   docker
    #   docker-compose

# The following packages are considered to be unsafe in a requirements file:
# setuptools
